Modelos y Repositorios cargados

+++ Models/Producto.cs
@@ -4,15 +4,7 @@ namespace tl2_tp8_2025_camip1.Models
     {
         public int IdProducto { get; set; }
         public string Descripcion { get; set; }
-        public double Precio { get; set; }
+        public int Precio { get; set; }
 
-        //Producto() { }
-
-        // Producto(int id, string desc, int precio)
-        // {
-        //     id_Producto = id;
-        //     descripcion = desc;
-        //     this.precio = precio;
-        // }
     }
 }
